LPD-20184 Move fixtures and pages to their correspondent tests directories
