+++ modules/test/playwright/tests/batch-planner/fixtures/dataMigrationCenterPages.ts
@@ -5,7 +5,7 @@
 
 import {test} from '@playwright/test';
 
-import {DataMigrationCenterPage} from '../pages/batch-planner/dataMigrationCenter.page';
+import {DataMigrationCenterPage} from '../pages/dataMigrationCenter.page';
 
 const dataMigrationCenterPagesTest = test.extend<{
 	dataMigrationCenterPage: DataMigrationCenterPage;
