+++ modules/test/playwright/tests/export-import-web/import.spec.ts
@@ -10,7 +10,7 @@ import * as path from 'path';
 
 import {apiHelpersTest} from '../../fixtures/apiHelpersTest';
 import {documentLibraryPagesTest} from '../../fixtures/documentLibraryPages.fixtures';
-import {exportImportPagesTest} from '../../fixtures/exportImportPages.fixtures';
+import {exportImportPagesTest} from './fixtures/exportImportPages.fixtures';
 import {loginTest} from '../../fixtures/loginTest';
 
 export const test = mergeTests(
