+++ modules/test/playwright/tests/headless-builder-impl/apiFilter.spec.ts
@@ -6,7 +6,7 @@
 import {APIResponse, expect as baseExpect, mergeTests} from '@playwright/test';
 
 import {apiHelpersTest} from '../../fixtures/apiHelpersTest';
-import {headlessBuilderTest} from '../../fixtures/headlessBuilderTest';
+import {headlessBuilderTest} from '../headless-builder-web/fixtures/headlessBuilderTest';
 import {loginTest} from '../../fixtures/loginTest';
 
 export const test = mergeTests(
