+++ modules/test/playwright/tests/headless-builder-web/fixtures/headlessBuilderPagesTest.ts
@@ -3,8 +3,8 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import {ApplicationPage} from '../pages/headless-builder-web/applicationPage';
-import {HeadlessBuilderPage} from '../pages/headless-builder-web/headlessBuilderPage';
+import {ApplicationPage} from '../pages/applicationPage';
+import {HeadlessBuilderPage} from '../pages/headlessBuilderPage';
 import {headlessBuilderTest} from './headlessBuilderTest';
 
 const headlessBuilderPagesTest = headlessBuilderTest.extend<{
