+++ modules/test/playwright/tests/headless-builder-web/pages/headlessBuilderPage.ts
@@ -5,7 +5,7 @@
 
 import {Locator, Page} from '@playwright/test';
 
-import {ApplicationsMenuPage} from '../product-navigation-applications-menu/ApplicationsMenuPage';
+import {ApplicationsMenuPage} from '../../../pages/product-navigation-applications-menu/ApplicationsMenuPage';
 
 export class HeadlessBuilderPage {
 	readonly addNewApplicationButton: Locator;
