+++ modules/test/playwright/tests/headless-builder-web/schemas.spec.ts
@@ -6,7 +6,7 @@
 import {expect, mergeTests} from '@playwright/test';
 
 import {apiHelpersTest} from '../../fixtures/apiHelpersTest';
-import {headlessBuilderPagesTest} from '../../fixtures/headlessBuilderPagesTest';
+import {headlessBuilderPagesTest} from './fixtures/headlessBuilderPagesTest';
 import {loginTest} from '../../fixtures/loginTest';
 
 export const test = mergeTests(
